Forums Ajust.

+++ extras/dictionary.c
@@ -141,10 +141,10 @@ dictionary * dictionary_new(size_t size)
 /*--------------------------------------------------------------------------*/
 void dictionary_del(dictionary * d)
 {
-    size_t  i ;
+    size_t  i;
 
     if (d==NULL) return ;
-    for (i=0 ; i<d->size ; i++) {
+    for (i=0 ; i < (size_t)d->size ; i++) {
         if (d->key[i]!=NULL)
             free(d->key[i]);
         if (d->val[i]!=NULL)
@@ -177,7 +177,7 @@ char * dictionary_get(dictionary * d, const char * key, char * def)
     size_t      i ;
 
     hash = dictionary_hash(key);
-    for (i=0 ; i<d->size ; i++) {
+    for (i=0 ; i < (size_t)d->size ; i++) {
         if (d->key[i]==NULL)
             continue ;
         /* Compare hash */
@@ -228,7 +228,7 @@ int dictionary_set(dictionary * d, const char * key, const char * val)
     hash = dictionary_hash(key) ;
     /* Find if value is already in dictionary */
     if (d->n>0) {
-        for (i=0 ; i<d->size ; i++) {
+        for (i=0 ; i<(size_t)d->size ; i++) {
             if (d->key[i]==NULL)
                 continue ;
             if (hash==d->hash[i]) { /* Same hash value */
@@ -263,7 +263,7 @@ int dictionary_set(dictionary * d, const char * key, const char * val)
        d->size. Because d->n < d->size this will necessarily
        terminate. */
     for (i=d->n ; d->key[i] ; ) {
-        if(++i == d->size) i = 0;
+        if(++i == (size_t)d->size) i = 0;
     }
     /* Copy key */
     d->key[i]  = xstrdup(key);
@@ -294,7 +294,7 @@ void dictionary_unset(dictionary * d, const char * key)
     }
 
     hash = dictionary_hash(key);
-    for (i=0 ; i<d->size ; i++) {
+    for (i=0 ; i<(size_t)d->size ; i++) {
         if (d->key[i]==NULL)
             continue ;
         /* Compare hash */
@@ -306,7 +306,7 @@ void dictionary_unset(dictionary * d, const char * key)
             }
         }
     }
-    if (i>=d->size)
+    if (i>=(size_t)d->size)
         /* Key not found */
         return ;
 
@@ -342,7 +342,7 @@ void dictionary_dump(dictionary * d, FILE * out)
         error( "empty dictionary" );
         return ;
     }
-    for (i=0 ; i<d->size ; i++) {
+    for (i=0 ; i<(size_t)d->size ; i++) {
         if (d->key[i]) {
             fprintf(out, "%20s\t[%s]\n",
                     d->key[i],
